feat: Add navigation to WorkoutSessionScreen from WorkoutExercisesScreen
This commit adds navigation functionality to the WorkoutSessionScreen from the WorkoutExercisesScreen. When the user taps on the "Начать тренировку" button, they will be directed to the WorkoutSessionScreen where they can start the workout session. This enhances the user experience by providing a seamless transition between screens and allows users to easily begin their workout.