feat: Add fetchClientCompletedExercises method to ApiRepository
This commit adds the fetchClientCompletedExercises method to the ApiRepository. This method allows fetching a list of completed exercises for a specific client and exercise ID. It makes an HTTP GET request to the API endpoint and returns a list of ClientWorkoutExercise objects. This functionality will be used to display the completed exercises for a client in the app.

Recent user commits:
- Add navigation to WorkoutSessionScreen from WorkoutExercisesScreen
- Update WorkoutExercisesScreen to display exercise equipment and sets description
- Update ClientWorkoutExercise model to use exerciseId instead of workoutExerciseId
- Fix TableCalendar for Android
- Add lazy loading for images in MainScreen
- Add internet permission to AndroidManifest.xml
- Update StoriesList to use ClientTrainerBloc for fetching client trainers
- Create README.md
- Update ChatListScreen and ChatScreen for chat functionality
- Add lazy loading for exercise images in WorkoutExercisesScreen