refactor: Update ClientTrainingPlansSection to use GestureDetector for training plan navigation
This commit updates the ClientTrainingPlansSection widget to use the GestureDetector widget for training plan navigation. Previously, the TrainingCard widget used the onTap property to navigate to the TrainingPlanScreen, but now the onTap property is passed down from the TrainingCard to the ClientTrainingPlansSection. This change improves code organization and allows for more flexibility in handling the navigation action. The user experience is enhanced by providing a consistent and intuitive way to navigate to the training plan screen.