feat: Add ClientListScreen for displaying a list of clients
This commit adds a new screen, ClientListScreen, which displays a list of clients. The screen includes a search bar for filtering clients, and each client is represented by a tile with their name and profile picture. The screen also includes a button to navigate to the ChatListScreen. This addition enhances the app's functionality by providing a dedicated screen for managing and interacting with clients.

Recent user commits:
- refactor: Update home screen to dynamically display appropriate screen based on user role
- Create home_trainer_screen.dart
- refactor: Add 'role' property to User model
- Rename: client_screen.dart -> home_client_screen.dart
- refactor: Update ProgressScreen to add onTap functionality to ProgressCard
- refactor: Update ProgressCard onTap functionality, HomeScreen and ProfileScreen icons and labels, and WorkoutExercisesScreen lazy loading
- refactor: Update HomeScreen and ProfileScreen icons and labels
- refactor: Update WorkoutExercisesScreen to use lazy loading for exercise images
- refactor: Update ProfileScreen to add 'Аккаунт' and 'Основное' section headers
- refactor: Update WorkoutExercisesScreen to use lazy loading for exercise images
- refactor: Update ProfileScreen to add 'Аккаунт' and 'Основное' section headers