feat: Add fetchClientsForTrainer method to ApiRepository
This commit adds the fetchClientsForTrainer method to the ApiRepository class. This method allows fetching a list of clients for a specific trainer by making an HTTP GET request to the appropriate API endpoint. The response is then parsed and converted into a list of User objects. If the request is successful, the list of clients is returned. Otherwise, an exception is thrown. This addition enhances the functionality of the app by providing the ability to retrieve and display clients for a trainer.