feat: Add fetchUserTrainingPlans method to ApiRepository
This commit adds the fetchUserTrainingPlans method to the ApiRepository class. This method allows fetching a list of training plans for a specific user by making an HTTP GET request to the appropriate API endpoint. The response is then parsed and converted into a list of TrainingPlan objects. If the request is successful, the list of training plans is returned. Otherwise, an exception is thrown. This addition enhances the functionality of the app by providing the ability to retrieve and display training plans for a user.